Add modules/torch_utils.py
